dataset/scripts/requirements.txt
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1
odfpy==1.4.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
